frontend code cleanup
