+++ client-interface/components/admin/certificates/hooks/index.ts
@@ -1,4 +1,2 @@
 export * from './useAIEvaluationProgress';
-export * from './useTierModal';
-export * from './useCertificateQualifications';
-export * from './useCertificateCanvas';
+export * from './useRecipientSelection';
